+++ accounts-manager/src/main/java/com/ghory/apps/accountsmanager/Application.java
@@ -1,20 +1,9 @@
-/*
- * Copyright 2012-2013 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+/**
+ * Main Application configuration class.
+ * 
  */
 
-package com.ghory.apps.accountsmanager.jetty;
+package com.ghory.apps.accountsmanager;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
@@ -24,10 +13,10 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 @EnableAutoConfiguration
 @ComponentScan
-public class SampleJettyApplication {
+public class Application {
 
 	public static void main(String[] args) throws Exception {
-		SpringApplication.run(SampleJettyApplication.class, args);
+		SpringApplication.run(Application.class, args);
 	}
 
 }
