+++ accounts-manager/src/main/java/com/ghory/apps/accountsmanager/service/HelloWorldService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.ghory.apps.accountsmanager.jetty.service;
+package com.ghory.apps.accountsmanager.service;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
