+++ accounts-manager/src/test/java/com/ghory/apps/accountsmanager/jetty/SampleJettyApplicationTests.java
@@ -28,6 +28,8 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.web.WebAppConfiguration;
 
+import com.ghory.apps.accountsmanager.Application;
+
 import static org.junit.Assert.assertEquals;
 
 /**
@@ -36,7 +38,7 @@ import static org.junit.Assert.assertEquals;
  * @author Dave Syer
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(classes = SampleJettyApplication.class)
+@SpringApplicationConfiguration(classes = Application.class)
 @WebAppConfiguration
 @IntegrationTest("server.port:0")
 @DirtiesContext
